Remove library calls that are now contained in setup.R and unnecessary text.
Update select_cfs source table following updated .csv from @cmothes.

Add code to import WBM for an AOI (e.g., watersupply watershed) to improve correspondence between water use area and water supply area.

Start working on a similar section for getting the climate for an AOI. Note, if we run the WBM model ourselves, we'll need this. If we continue to use gridded WBM, we probably can just continue to use the centroid temperature and precipitation for plots -- though it doesn't directly correspond to the AOI WBM...  It takes a bit of time to import the climate data using climateR. We should follow up on this at some point.

Remove the Asymptotic Regression model for use projections. This never performed well and we likely won't go back to it.